refactor(browser,bottomnav): default Content() for browser + bottomnav blocs
Move all four BrowserScreens (Browser, BrowserEditQuery, BrowserLanding,
BrowserRoot) and both BottomNavScreens into their public modules; each
Bloc interface now provides a default Content() that delegates to the
screen. BrowserRootBlocImpl keeps its custom Content override because it
wraps the screen in PlusHeaderContainer for the Modal presentation —
that behavior can't be expressed by the interface default.
