+++ client/bottomnav/impl/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/bottomnav/impl/BottomNavOrderBlocImpl.kt
@@ -1,14 +1,11 @@
 package com.plusmobileapps.chefmate.recipe.bottomnav.impl
 
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.di.AppScope
 import com.plusmobileapps.chefmate.getViewModel
 import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavOrderBloc
 import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavOrderBloc.Output
-import com.plusmobileapps.chefmate.recipe.bottomnav.impl.ui.BottomNavOrderScreen
 import com.plusmobileapps.metro.extensions.assistedfactory.ContributesAssistedFactory
 import dev.zacsweers.metro.Assisted
 import dev.zacsweers.metro.AssistedInject
@@ -41,10 +38,5 @@ class BottomNavOrderBlocImpl(
 
     override fun onBack() {
         output.onNext(Output.Back)
-    }
-
-    @Composable
-    override fun Content(modifier: Modifier) {
-        BottomNavOrderScreen(bloc = this, modifier = modifier)
     }
 }
