+++ client/bottomnav/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/bottomnav/BottomNavOrderBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.recipe.bottomnav
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.plusmobileapps.chefmate.BlocContext
 import com.plusmobileapps.chefmate.Consumer
 import com.plusmobileapps.chefmate.ui.ComposeScreen
@@ -9,6 +11,11 @@ import kotlinx.coroutines.flow.StateFlow
 
 interface BottomNavOrderBloc : ComposeScreen {
     val state: StateFlow<Model>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        BottomNavOrderScreen(bloc = this, modifier = modifier)
+    }
 
     fun onMove(from: Int, to: Int)
