+++ client/bottomnav/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/recipe/bottomnav/BottomNavOrderScreen.kt
@@ -1,4 +1,4 @@
-package com.plusmobileapps.chefmate.recipe.bottomnav.impl.ui
+package com.plusmobileapps.chefmate.recipe.bottomnav
 
 import androidx.compose.animation.AnimatedVisibility
 import androidx.compose.animation.core.animateDpAsState
@@ -46,8 +46,6 @@ import chefmate.client.bottomnav.public.generated.resources.Res
 import chefmate.client.bottomnav.public.generated.resources.bottom_nav_order_drag_handle_content_description
 import chefmate.client.bottomnav.public.generated.resources.bottom_nav_order_save
 import chefmate.client.bottomnav.public.generated.resources.bottom_nav_order_title
-import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavBloc
-import com.plusmobileapps.chefmate.recipe.bottomnav.BottomNavOrderBloc
 import com.plusmobileapps.chefmate.text.asTextData
 import com.plusmobileapps.chefmate.ui.components.PlusFloatingActionButton
 import com.plusmobileapps.chefmate.ui.components.PlusHeaderContainer
