+++ client/browser/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/browser/BrowserRootBloc.kt
@@ -1,5 +1,7 @@
 package com.plusmobileapps.chefmate.browser
 
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
 import com.arkivanov.decompose.router.stack.ChildStack
 import com.arkivanov.decompose.value.Value
 import com.plusmobileapps.chefmate.BlocContext
@@ -9,6 +11,11 @@ import com.plusmobileapps.chefmate.ui.ComposeScreen
 
 interface BrowserRootBloc : ComposeScreen {
     val routerState: Value<ChildStack<*, Child>>
+
+    @Composable
+    override fun Content(modifier: Modifier) {
+        BrowserRootScreen(bloc = this, modifier = modifier)
+    }
 
     fun navigateToUrl(url: String)
