+++ client/browser/public/src/commonMain/kotlin/com/plusmobileapps/chefmate/browser/BrowserRootScreen.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalSharedTransitionApi::class)
 
-package com.plusmobileapps.chefmate.browser.impl.ui
+package com.plusmobileapps.chefmate.browser
 
 import androidx.compose.animation.AnimatedContent
 import androidx.compose.animation.ExperimentalSharedTransitionApi
@@ -14,7 +14,6 @@ import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import com.arkivanov.decompose.extensions.compose.subscribeAsState
-import com.plusmobileapps.chefmate.browser.BrowserRootBloc
 
 @Composable
 fun BrowserRootScreen(bloc: BrowserRootBloc, modifier: Modifier = Modifier) {
